58 person credit refactor (#59)
* feat #58 Popular person card 텍스트 패딩 및 높이 조정

* feat #58 PersonCredit 적용

* feat #58 api 동시 호출 되도록 수정

* feat #58 person id int type 변경

* chore #58 줄 정리 수정

* feat #58 credit mapping 로직 개선

* feat #58 api param 버그 수정

* feat #58 PersonID 적용

* feat #58 코드 복구

+++ data/src/main/java/com/dcs/data/remote/datasource/PersonRemoteDataSource.kt
@@ -11,7 +11,7 @@ interface PersonRemoteDataSource {
         language: String,
     ): Result<GetPopularPeopleResponse>
 
-    suspend fun getPersonDetail(personId: Long): Result<GetPersonDetailResponse>
+    suspend fun getPersonDetail(personId: Int): Result<GetPersonDetailResponse>
 
     suspend fun getSearchPerson(personName: String): Result<GetSearchPersonResponse>
 }
